[ExportVerilog] Add parallelization assertion; NFC
Add a flag to `SharedEmitterState` that indicates whether processing
currently occurs in a parallel region. Parts of the emission code that
have to rely on sequential execution can then assert on this flag to
make sure they actually are executed in a sequential region.

This requires adding a reference from the current `VerilogEmitterState`
to the global `SharedEmitterState`, which tracks various aspects of the
emission, including the parallelization flag. Thanks to C++ being an
archaic language, this also means moving the entire `SharedEmitterState`
declaration up in the file.

This is only intended to provide a safety feature, and should not change
functionality at all. Most diff lines are due to code moving about.
